Remove MACH_P1 code
As part of the overall clean up of the stock source, this is the first of many
board defines that will be removed.

MACH_P1 is easy, as it is only affecting tv out components and one include
file. The gpio-p1.h is no longer needed as of: d547812b3e7f2daf7f061d7cbbfba7a2de57e4c7

Signed-off-by: Bryan Hundven <bryanhundven@gmail.com>