[SCRUB] Part deux...
Ran:
    git diff v2.6.35.7..android-samsung-2.6.35-galaxys4g-upgrade > ../scrub-deux.patch
    git checkout -b android-samsung-2.6.35-galaxys4g-scrub-deux v2.6.35.7
    patch -p1 < ../scrub-deux.patch

Which brings us to this commit.

This is essentially a history rewrite. Since the changes rewritten are
my changes and I've shown how we got from here to there, I'm not
opposed.

Previously, we had the upstream kernel history from 2.6.12 to 2.6.35.7,
then the initial port of the stock sgs4g source. This will make it hard
to merge to a newer kernel version in the future.

Also, now that everything it currently takes to get sgs4g to boot and
work properly is in one commit at the HEAD. This will also make it easier
to find commits to cherry pick from the aosp samsung kernel to provide
the original commits to some of the code that is already applied to our
tree.

Again, this is the scrubbing process...

Signed-off-by: Bryan Hundven <bryanhundven@gmail.com>